api-gateway/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
fastapi==0.114.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
pydantic==2.9.1
pydantic_core==2.23.3
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
python-jose  # For JWT
passlib[bcrypt] 